add abstactions to class
